pyproject.toml
[build-system]
requires = ["setuptools>=68,<82"]
build-backend = "setuptools.build_meta"

[project]
name = "football-perspectives"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "numpy>=1.26,<2.0",
    "scipy>=1.11,<1.14",
    "opencv-python>=4.10,<4.11",
    "scenedetect[opencv]>=0.6.4",
    "ultralytics>=8.0",
    "tqdm>=4.66",
    "supervision>=0.20",
    "transformers>=4.40",
    "scikit-learn>=1.4",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "fastapi>=0.111,<0.115",
    "uvicorn[standard]>=0.30,<0.35",
    "python-multipart>=0.0.9",
    "openmim>=0.3",
    "mmengine>=0.10,<0.11",
    "mmcv>=2.1,<2.2",
    "mmpretrain>=1.0,<1.3",
    "mmpose>=1.3,<1.4",
    "mmdet>=3.3,<3.4",
    "xtcocotools==1.14.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]
tvcalib = [
    "kornia>=0.7,<0.8",
    "SoccerNet>=0.1.56",
]
hmr = [
    "smplx>=0.1.28",
    "roma>=1.4",
    "pytorch_lightning>=2.0",
    "hydra-core>=1.3",
    "hydra-zen",
    "hydra_colorlog",
    "timm>=0.9",
    "colorlog",
    "imageio>=2.34",
    "av>=11",
    "ffmpeg-python",
    "einops",
    "chumpy",
    "wis3d>=1.0",
    "ultralytics>=8.2",
    "cython_bbox",
    "lapx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
